requirements.txt
# Core ML/DL Libraries
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# OCR and Barcode
pytesseract>=0.3.10
pyzbar>=0.1.9

# Image Processing and Augmentation
albumentations>=1.3.0

# PDF Processing
PyMuPDF>=1.23.0

# Data Science
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Web Framework
streamlit>=1.28.0

# API Requests (for mock validation)
requests>=2.31.0

# Utilities
tqdm>=4.66.0

